src/requirements.txt
assertpy==1.1
coverage==6.5.0
dataclasses-json==0.5.7
dependency-injector==4.40.0
elemental-cms==1.1.16
flask-csrf==0.9.2
flask-wtf==1.0.1
gunicorn==20.1.0
pip-chill==1.0.1
pytest==7.2.0
unittest2==1.1.0
